api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pyjwt==2.7.0
passlib[bcrypt]==1.7.4
pymysql==1.0.2
cryptography==41.0.0
pydantic[email]==1.10.8
pyjwt
requests